+++ R/variance_bootstrap.R
@@ -241,7 +241,8 @@ variance_bootstrap <- function(
   subset,
   parallel = "no",
   ncpus = 1L,
-  subset_env = parent.frame()
+  subset_env = parent.frame(),
+  trim = 1
 ) {
   data <- fit$data
   int_names <- names(interventions)
@@ -324,7 +325,8 @@ variance_bootstrap <- function(
               ipw_point_b <- compute_ipw_contrast_point(
                 fit_b,
                 interventions,
-                target_idx_b
+                target_idx_b,
+                trim = trim
               )
               ipw_point_b$mu_hat
             },
@@ -362,7 +364,8 @@ variance_bootstrap <- function(
               aipw_point_b <- compute_aipw_contrast_point(
                 fit_b,
                 interventions,
-                target_idx_b
+                target_idx_b,
+                trim = trim
               )
               aipw_point_b$mu_hat
             },